pyproject.toml
[project]
name = "pareto-optimization"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.2.1",
    "coco-experiment>=2.8.1",
    "cocopp>=2.7.3",
    "h5py>=3.15.1",
    "hdbscan>=0.8.40",
    "ipykernel>=6.29.5",
    "matplotlib>=3.10.1",
    "nbformat>=5.10.4",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pip>=25.1.1",
    "plotly[express]>=6.0.1",
    "pydantic>=2.11.5",
    "pykrige>=1.7.2",
    "pymoo>=0.6.1.3",
    "requests>=2.32.3",
    "safetensors>=0.7.0",
    "scikit-learn>=1.6.1",
    "seaborn>=0.13.2",
    "sklearn-som>=1.1.0",
    "torch>=2.7.0",
    "tqdm>=4.67.1",
    "umap-learn>=0.5.9.post2",
    "wandb>=0.20.1",
]


[dependency-groups]
dev = [
    "poethepoet>=0.40.0",
]



[tool.poe.tasks]

# ====================================================================================
# 1. Data Pipeline
# ====================================================================================

[tool.poe.tasks.data-gen]
shell = "python -m src.cli.generating.generate_dataset --function-id $fid"
help = "GENERATE: Create synthetic Pareto data. Usage: --fid 5"
  [[tool.poe.tasks.data-gen.args]]
  name = "fid"
  options = ["--fid"]
  default = "5"

[tool.poe.tasks.data-viz]
shell = "python -m src.cli.visualizing.visualize_dataset --dataset-name $data"
help = "DATA VIZ: View generated data space. Usage: --data cocoex_f5"
  [[tool.poe.tasks.data-viz.args]]
  name = "data"
  options = ["--data"]
  default = "cocoex_f5"

# ====================================================================================
# 2. Training Pipeline
# ====================================================================================

[tool.poe.tasks.train-inv]
shell = "python -m src.cli.training.train_inverse_model_standard --estimator $est --dataset-name $data"
help = "TRAIN INVERSE: Usage: uv run poe train-inv --est mdn --data cocoex_f5"
  [[tool.poe.tasks.train-inv.args]]
  name = "est"
  options = ["--est"]
  default = "mdn"
  choices = ["mdn", "cvae", "inn", "gan"]
  [[tool.poe.tasks.train-inv.args]]
  name = "data"
  options = ["--data"]
  default = "cocoex_f5"

[tool.poe.tasks.train-fwd]
shell = "python -m src.cli.training.train_forward_model standard --estimator $est --dataset-name $data"
help = "TRAIN FORWARD: Usage: uv run poe train-fwd --est mdn --data cocoex_f5"
  [[tool.poe.tasks.train-fwd.args]]
  name = "est"
  options = ["--est"]
  default = "mdn"
  choices = ["mdn", "mlp", "rf"]
  [[tool.poe.tasks.train-fwd.args]]
  name = "data"
  options = ["--data"]
  default = "cocoex_f5"


# ====================================================================================
# 3. Generation & Candidates
# ====================================================================================

[tool.poe.tasks.gen-candidates]
cmd = "python -m src.cli.generating.generate_candidates"
help = "CANDIDATES: Generate decisions for target objectives."

[tool.poe.tasks.compare-candidates]
cmd = "python -m src.cli.evaluating.compare_inverse_models_candidates_generation"
help = "COMPARE: Benchmarks candidate generation across models."

# ====================================================================================
# 4. Evaluation (The Auditor)
# ====================================================================================

[tool.poe.tasks.diagnose]
cmd = "python -m src.cli.evaluating.diagnose_inverse_models"
help = "DIAGNOSE: Run model diagnostics (PIT, MACE, Diversity) and save JSON."

[tool.poe.tasks.viz-diagnostics]
cmd = "python -m src.cli.evaluating.visualize_diagnostics"
help = "REPORT: Visualize historical diagnostics from the auditor."


# ====================================================================================
# 5. Performance Visualization (Specific Lenses)
# ====================================================================================

[tool.poe.tasks.viz-perf-inv]
shell = "python -m src.cli.evaluating.visualize_model_performance"
help = "VIZ PERF INVERSE: Detailed plots for a run. Usage: uv run poe"
  